bcb_coap: bcb_tc_default: Functionality to recover after OCP/OTP activated.
A configurable delay is added to close the switch after OTP/OCP is activated.
For AC, this delay will be a multiple of half sine period.
For DC, this will be a multiple of BCB_TRIP_CURVE_DEFAULT_TRANSIENT_WORK_TIMEOUT
which is configured via KConfig.
The delay can also be configured via CoAP with a POST request to the following
URI.
coap://<host>:<port>/tc_default?rd=<delay>

Fixes #5

Signed-off-by: Kasun Hewage <kasun.ch@gmail.com>