Use separate interrupt handlers for switch open and close events
Due to interrupt handler serving latency, distinguishing close (on)
and open (off) events may not be possible if they occur within very
short duration.
Therefore, separate interrupt handlers are used for rising and
falling edge events.
As this cannot be done with a single GPIO, mirrored pin of the
ON_OFF_STATUS signal is used for the falling edge event.

Fixes #9.

Signed-off-by: Kasun Hewage <kasun.ch@gmail.com>